# perf: Improve output speed by caching author_list