--- a/src/afterfact.rs
+++ b/src/afterfact.rs
@@ -423,14 +423,10 @@ fn emit_csv<W: std::io::Write>(
             // 各種集計作業
             if !output_option.no_summary {
                 let level_suffix = get_level_suffix(detect_info.level.as_str());
-                let author_list = match author_list_cache.get(&detect_info.rulepath) {
-                    None => {
-                        let x = extract_author_name(&detect_info.rulepath, stored_static);
-                        author_list_cache.insert(detect_info.rulepath.clone(), x.clone());
-                        x
-                    }
-                    Some(x) => x.clone(),
-                };
+                let author_list = author_list_cache
+                    .entry(detect_info.rulepath.clone())
+                    .or_insert_with(|| extract_author_name(&detect_info.rulepath, stored_static))
+                    .clone();
                 let author_str = author_list.iter().join(", ");
                 if !detected_rule_files.contains(&detect_info.rulepath) {
                     detected_rule_files.insert(detect_info.rulepath.to_owned());