@YamatoSecurity 2023-06-08
@fukusuket Thank you so much for the quick PR!
It did improve but still not as fast as 2.5.1.
I updated the benchmarks here: https://docs.google.com/spreadsheets/d/1XCEkJHYA5N8b7O2KldGp7q0C7_F1HFnibz7HwI90-s0/edit#gid=0
On 32GB of data, output processing time is:
```
2.5.1: 11 min
main branch: 1 hour 7 min
this PR: 37 min
```
By the way, is the rule author list only needed for outputting author names next to rules in the HTML report?
If this is the case, can we just compute that information when `-H` is specified?