@fukusuket 2023-06-08
@YamatoSecurity 
I have one question! 
ExtraFieldInfo was added in version 2.6.0, so I think you'll need to use the same profile for an accurate comparison with 2.5.1.
Is there a difference in performance even when using the same profile?