@YamatoSecurity 2023-06-08
@fukusuket That is a good point!
Yes, `-p minimal` is very similar to (same as?) 2.5.1's default output.
I will re-run my benchmark