@fukusuket 2023-06-08
I checked with the data of all-evtx.tgz (6.1GB). The results are as follows.
It seems that it is a little slower when output?🤔 I will investigate where improving the output processing as much as I can.

|Ver |Elasped Time| Rule Parse Time | Analysis Time | Output Time| Memory peak |
|-- | ---- | -- | ---- | ---- |---- | 
|2.5.1 | 00:06:29.145 | 00:00:01.100 | 00:06:20.710 | 00:00:07.333 | 4.0 GiB |
|This PR| 00:06:32.009 | 00:00:01.106 | 00:06:22.602 | 00:00:08.300 | 4.0 GiB |