@YamatoSecurity 2023-06-09
@fukusuket 悲報: Unfortunately there is not a big improvement for my larger dataset (32GB). Output time does decrease from 34 min 30 seconds to 34 min 10 sec, but 2.5.1 is only around 11 min.
Could you see if there is anything else that might be causing the slower times?