@YamatoSecurity 2023-06-10
@hitenkoku is this PR is included in https://github.com/Yamato-Security/hayabusa/pull/1089 should we close it?