@fukusuket 2023-06-10
Yes, it's included in #1088, so I'll close it! :)